+++ package.json
@@ -3,28 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.17",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.25.0",
-    "lodash": "^4.17.21",
-    "pokemon-images": "^2.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-paginate": "^8.1.0",
-    "react-poke-sprites": "^0.0.8",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
-    "react-spinners": "^0.11.0",
     "redux": "^4.1.2",
-    "redux-devtools-extension": "^2.13.9",
+    "redux-actions": "^2.6.5",
     "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.1",
     "styled-components": "^5.3.3",
     "web-vitals": "^2.1.4"
   },
