--- src/components/Player/Music.jsx
@@ -1,25 +1,0 @@
-import {useState, useEffect} from 'react';
-import Player from './Player';
-
-function App() {
-
-  const [songs] = useState([
-    {
-      src: "./Music/abertura.mp3"
-    },
-  ]);
-
-  const [currentSongIndex, setCurrentSongIndex] = useState(0);
-  
-  return (
-    <>
-      <Player 
-        currentSongIndex={currentSongIndex} 
-        setCurrentSongIndex={setCurrentSongIndex} 
-        songs={songs}
-      />
-    </>
-  );
-}
-
-export default App;
