--- src/components/Player/Player.js
@@ -1,51 +1,0 @@
-import React, {useState, useRef, useEffect} from 'react'
-import Controls from './Controls';
-import * as S from "../Player/styles";
-
-function Player(props) {
-    const audioEl = useRef(null);
-    const [isPlaying, setIsPlaying] = useState(false);
-
-    useEffect(() => {
-        if (isPlaying) {
-            audioEl.current.play();
-        } else {
-            audioEl.current.pause();
-        }
-    });
-
-    const SkipSong = (forwards = true) => {
-        if (forwards) {
-            props.setCurrentSongIndex(() => {
-                let temp = props.currentSongIndex;
-                temp++;
-
-                if (temp > props.songs.length - 1) {
-                    temp = 0;
-                }
-
-                return temp;
-            });
-        } else {
-            props.setCurrentSongIndex(() => {
-                let temp = props.currentSongIndex;
-                temp--;
-
-                if (temp < 0) {
-                    temp = props.songs.length - 1;
-                }
-
-                return temp;
-            });
-        }
-    }
-
-    return (
-        <S.Player>
-            <audio src={props.songs[props.currentSongIndex].src} ref={audioEl}></audio>
-            <Controls isPlaying={isPlaying} setIsPlaying={setIsPlaying} SkipSong={SkipSong} />
-        </S.Player>
-    )
-}
-
-export default Player;
