+++ src/pages/Carrinho/Carrinho.style.js
@@ -29,7 +29,7 @@ export const Div = styled.div`
 `;
 
 export const TopDiv = styled.div`
-    width: 920px;
+    width: 910px;
     height: 20px;
     left: 239px;
     top: 115px;
@@ -40,7 +40,7 @@ export const TopDiv = styled.div`
 `;
 
 export const ItemDiv = styled.div`
-    width: 912px;
+    width: 920px;
     height: 400px;
     left: 239px;
     top: 115px;
@@ -78,7 +78,7 @@ overflow: scroll;
 
 export const Item = styled.div`
 	flex: 1;
-	margin: 3px;
+	margin: 5px;
 	text-align: center;
 	font-size: 1.5em;
     font-family: Lato;
@@ -100,7 +100,6 @@ text-align: right;
 `;
 
 export const TituloCenter = styled(TituloCol)`
-margin-left: 32px;
 text-align: center;
 `;
