--- src/pages/Home/style.home.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-   img {
-    width: 100%;
-    height: 620px;
-   }
-`;
