--- src/redux/action/CounterAction.js
@@ -1,11 +1,0 @@
-import TYPES from "../types/types";
-
-export const increaseAction = (payload) => ({
-  type: TYPES.increase,
-  payload: payload,
-});
-
-export const decreaseAction = (payload) => ({
-  type: "DECREASE",
-  payload: payload,
-});
