+++ src/redux/actions/CounterAction.js
@@ -1,0 +1,11 @@
+import TYPES from "../types/types";
+
+export const increaseAction = (payload) => ({
+  type: TYPES.increase,
+  payload: payload,
+});
+
+export const decreaseAction = (payload) => ({
+  type: "DECREASE",
+  payload: payload,
+});
