--- src/redux/reducers/RootReducers.js
@@ -1,12 +1,0 @@
-import {combineReducers} from "redux";
-import PokemonListReducer from "./PokemonListReducer";
-import PokemonMultipleReducer from "./PokemonMultipleReducer";
-import CounterReducer from "./Counter.reducer";
-
-const RootReducer = combineReducers({
-  PokemonList: PokemonListReducer,
-  Pokemon: PokemonMultipleReducer,
-  counter: CounterReducer,
-});
-
-export default RootReducer;
